spiffmain skeleton linkage
