+++ .gitignore
@@ -1,3 +1,4 @@
 .sw?
 .*.sw?
 /build
+/tags
