+++ cli/CMakeLists.txt
@@ -4,3 +4,4 @@ project(spffl_cli)
 
 add_subdirectory(gcd)
 add_subdirectory(psdes)
+add_subdirectory(spiff)
