+++ cli/gcd/CMakeLists.txt
@@ -1,13 +1,13 @@
 cmake_minimum_required (VERSION 3.24)
 
-project(spffl_cli)
+project(cli_gcd)
 
 add_executable(cli_gcd cli_gcd.cpp)
 
-target_include_directories(cli_gcd
+target_include_directories(${PROJECT_NAME}
   PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/../../src/intmath
 )
-target_link_libraries(cli_gcd PRIVATE
+target_link_libraries(${PROJECT_NAME} PRIVATE
     intmath
 )
