+++ cli/psdes/CMakeLists.txt
@@ -1,13 +1,17 @@
 cmake_minimum_required (VERSION 3.24)
 
-project(spffl_cli)
+project(cli_psdes)
 
-add_executable(cli_psdes cli_psdes.cpp)
+add_executable(
+  cli_psdes
+  cli_psdes.cpp
+)
 
-target_include_directories(cli_psdes
+target_include_directories(${PROJECT_NAME}
   PRIVATE
     ${CMAKE_CURRENT_SOURCE_DIR}/../../src/bitrand
 )
-target_link_libraries(cli_psdes PRIVATE
+target_link_libraries(${PROJECT_NAME}
+  PRIVATE
     bitrand
 )
