+++ src/CMakeLists.txt
@@ -10,3 +10,4 @@ add_subdirectory(polynomials)
 add_subdirectory(rationals)
 add_subdirectory(random)
 add_subdirectory(factorization)
+add_subdirectory(cliparser)
