+++ src/cliparser/CMakeLists.txt
@@ -1,19 +1,17 @@
 cmake_minimum_required (VERSION 3.24)
 
-# This directory is headers-only
-
-project(containers)
+project(cliparser)
 
 add_library(${PROJECT_NAME}
-  containers.cpp
+  cmdinterp.cpp
 )
 
 target_include_directories(${PROJECT_NAME}
   PUBLIC ${PROJECT_SOURCE_DIR}/include
   PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../src/base
-  PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../src/intmath
+  PRIVATE ${CMAKE_CURRENT_SOURCE_DIR}/../../src/containers
 )
 target_link_libraries(${PROJECT_NAME} PRIVATE
     base
-    intmath
+    containers
 )
